requirements.txt
pyqt5
dss-python
numpy